requirements.txt
hdx-python-api==6.3.2
python-slugify==8.0.4
beautifulsoup4==4.12.3
html2text==2024.2.26
lxml==5.2.2